+++ demo/js/directives.js
@@ -147,7 +147,7 @@
 				children: [
 					{
 						node: {
-							filterable: true,
+							filterType: Lui.Directives.FilterTypeEnum.TEXT,
 							fixed: true,
 							label: "id",
 							width: 20,
@@ -162,12 +162,12 @@
 					},
 					{
 						node: {
-							filterable: true,
+							filterType: Lui.Directives.FilterTypeEnum.SELECT,
 							fixed: false,
 							label: "name",
 							width: 20,
+							getFilterValue: function (someone) { return someone.name; },
 							getValue: function (someone) { return "<span>" + someone.name + "</span>"; },
-							getOrderByValue: function (someone) { return someone.name; },
 							colspan: null,
 							hidden: false,
 							rowspan: null,
@@ -177,7 +177,7 @@
 					},
 					{
 						node: {
-							filterable: true,
+							filterType: Lui.Directives.FilterTypeEnum.TEXT,
 							fixed: false,
 							label: "adress",
 							width: 20,
@@ -192,7 +192,7 @@
 					},
 					{
 						node: {
-							filterable: true,
+							filterType: Lui.Directives.FilterTypeEnum.TEXT,
 							fixed: false,
 							label: "contacts",
 							width: 20,
@@ -206,7 +206,7 @@
 						children: [
 							{
 								node: {
-									filterable: true,
+									filterType: Lui.Directives.FilterTypeEnum.MULTISELECT,
 									fixed: false,
 									label: "phone",
 									width: 20,
@@ -221,7 +221,7 @@
 							},
 							{
 								node: {
-									filterable: true,
+									filterType: Lui.Directives.FilterTypeEnum.TEXT,
 									fixed: false,
 									label: "mail",
 									width: 20,