+++ ts/table-grid/table-grid.classes.ts
@@ -10,15 +10,16 @@ module Lui.Directives.TableGrid {
 
 	export class Header {
 		public label: string;
-		public filterable: boolean;
+		public filterType: string;
 		public hidden: boolean;
 		public width: number;
-		public fixed: boolean;
+		public fixed: string;
 		public colspan: number;
 		public rowspan: number;
 		public textAlign: string;
 		public getValue: (object: any) => string;
 		public getOrderByValue: (object: any) => any;
+		public getFilterValue: (object: any) => any;
 	}
 
 	export class BrowseResult {