+++ ts/table-grid/table-grid.controller.ts
@@ -4,6 +4,13 @@ module Lui.Directives {
 
 	"use strict";
 
+	export class FilterTypeEnum {
+		public static NONE = "none";
+		public static TEXT = "text";
+		public static SELECT = "select";
+		public static MULTISELECT = "multiselect";
+	}
+
 	export class LuidTableGridController {
 		public static IID: string = "luidTableGridController";
 		public static $inject: Array<string> = ["$filter", "$scope", "$translate"];
@@ -27,25 +34,17 @@ module Lui.Directives {
 				if (result.tree.children.length) {
 					result.subDepth++;
 				} else {
-					if (result.tree.node.fixed) {
-						$scope.fixedRowDefinition.push(result.tree.node);
-					} else {
-						$scope.scrollableRowDefinition.push(result.tree.node);
-					}
+					$scope.colDefinitions.push(result.tree.node);
 				}
 
 				if (result.tree.node) {
 					result.tree.node.rowspan = maxDepth - result.depth - result.subDepth;
 					result.tree.node.colspan = result.subChildren;
-					if (!result.tree.children.length && !result.tree.node.filterable) {
+					if (!result.tree.children.length && result.tree.node.filterType === FilterTypeEnum.NONE) {
 						result.tree.node.rowspan++;
 					}
 
-					if (result.tree.node.fixed) {
-						$scope.fixedHeaderRows[result.depth] ? $scope.fixedHeaderRows[result.depth].push(result.tree.node) : $scope.fixedHeaderRows[result.depth] = [result.tree.node];
-					} else {
-						$scope.scrollableHeaderRows[result.depth] ? $scope.scrollableHeaderRows[result.depth].push(result.tree.node) : $scope.scrollableHeaderRows[result.depth] = [result.tree.node];
-					}
+					$scope.headerRows[result.depth] ? $scope.headerRows[result.depth].push(result.tree.node) : $scope.headerRows[result.depth] = [result.tree.node];
 				}
 				return result;
 			};
@@ -58,43 +57,60 @@ module Lui.Directives {
 				return depth + 1;
 			};
 
-			let init = () => {
+			let initFilter = () => {
+				$scope.filters = [];
+				_.each($scope.datas, (row: any) => {
+					_.each($scope.colDefinitions, (header: TableGrid.Header, index: number) => {
+						if (!$scope.filters[index]) {
+							$scope.filters[index] = { header: header, selectValues: [], currentValues: [] };
+						}
+						if (header.filterType === FilterTypeEnum.SELECT
+								|| header.filterType === FilterTypeEnum.MULTISELECT) {
+							let value = header.getValue(row);
+							if (!!header.getFilterValue) {
+								value = header.getFilterValue(row);
+							}
 
-				$scope.fixedHeaderRows = [];
-				$scope.fixedRowDefinition = [];
-				$scope.scrollableHeaderRows = [];
-				$scope.scrollableRowDefinition = [];
+							let valuesToCheck = value.split("|");
+							_.each(valuesToCheck, (val: string) => {
+								if (!_.contains($scope.filters[index].selectValues, val)) {
+									$scope.filters[index].selectValues.push(val);
+								}
+							});
+						}
+					});
+
+				});
+			};
+
+			let init = () => {
+				$scope.FilterTypeEnum = FilterTypeEnum;
+				$scope.headerRows = [];
+				$scope.bodyRows = [];
+				$scope.colDefinitions = [];
+				$scope.allChecked = {value: false};
 
 				maxDepth = getTreeDepth($scope.header);
 
 				browse({ depth: 0, subChildren: 0, subDepth: 0, tree: $scope.header });
 
-				let diff = $scope.fixedHeaderRows.length - $scope.scrollableHeaderRows.length;
-				if (diff > 0) {
-					for (let i = 1; i <= diff; i++) {
-						$scope.scrollableHeaderRows.push([]);
-					}
-				} else if (diff < 0) {
-					for (let i = 1; i <= -diff; i++) {
-						$scope.fixedHeaderRows.push([]);
-					}
-				}
 
 				$scope.selected = { orderBy: null, reverse: false };
 
-				$scope.leftFilters = [];
-				$scope.rightFilters = [];
+				initFilter();
 			};
 
-			let updateFilteredAndOrderedRows = () => {
+			$scope.updateFilteredAndOrderedRows = () => {
 				let temp = _.chain($scope.datas)
 					.filter((row: any) => {
 						let result = true;
-						let filters = $scope.leftFilters.concat($scope.rightFilters);
-						filters.forEach((filter: { header: TableGrid.Header, value: string }) => {
-							if (filter.header && filter.value && filter.value !== "") {
+						$scope.filters.forEach((filter: { header: TableGrid.Header, selectValues: string[], currentValues: string[] }) => {
+							if (filter.header
+									&& !!filter.currentValues[0]
+									&& filter.currentValues[0] !== "") {
 								let prop = (filter.header.getValue(row) + "").toLowerCase();
-								if (prop.indexOf(filter.value.toLowerCase()) === -1) {
+								let containsProp = _.some(filter.currentValues, (value: string) => { return prop.indexOf(value.toLowerCase()) !== -1; });
+								if (!containsProp) {
 									result = false;
 								}
 							}
@@ -103,51 +119,70 @@ module Lui.Directives {
 					});
 				if ($scope.selected && $scope.selected.orderBy) {
 					temp = temp.sortBy((row: any) => {
-						return $scope.selected.orderBy.getOrderByValue(row);
+						let orderByValue = $scope.selected.orderBy.getValue(row);
+						if ( $scope.selected.orderBy.getOrderByValue != null) {
+							orderByValue = $scope.selected.orderBy.getOrderByValue(row);
+						}
+						return orderByValue;
 					});
 				}
 				let filteredAndOrderedRows = temp.value();
 				$scope.filteredAndOrderedRows = $scope.selected.reverse ? filteredAndOrderedRows.reverse() : filteredAndOrderedRows;
+
 				$scope.updateVirtualScroll();
 			};
 
-			// orderBys and filterBys
-
-			$scope.updateFilterBy = (header: TableGrid.Header, index: number) => {
-				let value = header.fixed ? $scope.leftFilters[index].value : $scope.rightFilters[index].value;
-				if (!header) { return; }
-				if (!header.filterable) { return; }
-				if (value === null || value === "") {
-					header.fixed ? $scope.leftFilters[index] = { header: null, value: "" } : $scope.rightFilters[index] = { header: null, value: "" };
-				}
-				header.fixed ? $scope.leftFilters[index] = { header: header, value: value } : $scope.rightFilters[index] = { header: header, value: value };
-
-				updateFilteredAndOrderedRows();
-			};
-
 			$scope.updateOrderBy = (header: TableGrid.Header) => {
-				if (header.getOrderByValue != null) {
-					if (header === $scope.selected.orderBy) {
-						if ($scope.selected.reverse) {
-							$scope.selected.orderBy = null;
-							$scope.selected.reverse = false;
-						} else {
-							$scope.selected.reverse = true;
-						}
-					} else {
-						$scope.selected.orderBy = header;
+				if (header === $scope.selected.orderBy) {
+					if ($scope.selected.reverse) {
+						$scope.selected.orderBy = null;
 						$scope.selected.reverse = false;
+					} else {
+						$scope.selected.reverse = true;
 					}
+				} else {
+					$scope.selected.orderBy = header;
+					$scope.selected.reverse = false;
 				}
 
-				updateFilteredAndOrderedRows();
+				$scope.updateFilteredAndOrderedRows();
 			};
 
-			// strip html for display in title attribute
-			$scope.stripHtml = (html) => {
-				let tmp = document.createElement("DIV");
-				tmp.innerHTML = html;
-				return tmp.textContent || tmp.innerText || "";
+			$scope.onMasterCheckBoxChange = () => {
+				if (_.some($scope.filteredAndOrderedRows, (row: any) => { return !row.isChecked; })) {
+					_.each($scope.filteredAndOrderedRows, (row: any) => { row.isChecked = true; });
+				} else {
+					_.each($scope.filteredAndOrderedRows, (row: any) => { row.isChecked = false; });
+				}
+			};
+
+			$scope.onCheckBoxChange = () => {
+				if (_.some($scope.filteredAndOrderedRows, (row: any) => { return !row.isChecked; })) {
+					$scope.allChecked.value = false;
+				} else {
+					$scope.allChecked.value = true;
+				}
+			};
+
+			$scope.getCheckboxState = () => {
+				let selectedCheckboxesCount = _.where($scope.filteredAndOrderedRows, { isChecked: true }).length;
+				if (selectedCheckboxesCount === 0) {
+					return "";
+				}
+				if (selectedCheckboxesCount === $scope.filteredAndOrderedRows.length) {
+					return "checked";
+				}
+				if (selectedCheckboxesCount < $scope.filteredAndOrderedRows.length) {
+					return "partial";
+				}
+				return "";
+			};
+
+			$scope.clearSelect = ($select: any, $index: number, $event: any) => {
+				$event.stopPropagation();
+				$select.selected = undefined;
+				$scope.filters[$index].currentValues[0] = "";
+				$scope.updateFilteredAndOrderedRows();
 			};
 
 			// playing init