+++ ts/table-grid/table-grid.lang.ts
@@ -1,0 +1,22 @@
+angular.module("lui.translates.tablegrid").config(["$translateProvider", function ($translateProvider: angular.translate.ITranslateProvider): void {
+		$translateProvider.translations("en", {
+			"SELECT_ITEM": "Select an item",
+			"SELECT_ITEMS": "Select items",
+		});
+		$translateProvider.translations("de", {
+
+		});
+		$translateProvider.translations("es", {
+
+		});
+		$translateProvider.translations("fr", {
+			"SELECT_ITEM": "Séléctionnez un élément",
+			"SELECT_ITEMS": "Séléctionnez des éléments",
+		});
+		$translateProvider.translations("it", {
+
+		});
+		$translateProvider.translations("nl", {
+
+		});
+}]);