+++ ts/table-grid/table-grid.scope.ts
@@ -5,23 +5,35 @@ module Lui.Directives {
 	"use strict";
 
 	export interface IDataGridScope extends angular.IScope {
-		canvasHeight: any;
+		allChecked: any;
+		canvasHeight: number;
 		datas: any[];
 		filteredAndOrderedRows: any[];
-		fixedHeaderRows: TableGrid.Header[][];
+		headerRows: TableGrid.Header[][];
+		lockedWidth: number;
+		bodyRows: TableGrid.Header[][];
+		colDefinitions: TableGrid.Header[];
+		FilterTypeEnum: FilterTypeEnum;
+
 		fixedRowDefinition: TableGrid.Header[];
 		header: TableGrid.Tree;
-		leftFilters: {header: TableGrid.Header, value: string}[];
-		rightFilters: {header: TableGrid.Header, value: string}[];
-		scrollableHeaderRows: TableGrid.Header[][];
+		filters: {header: TableGrid.Header, selectValues: string[], currentValues: string[]}[];
 		scrollableRowDefinition: TableGrid.Header[];
+		selectable: boolean;
 		selected: { orderBy: TableGrid.Header, reverse: boolean };
 		visibleRows: any[];
 
-		stripHtml(html: string): string;
-		updateFilterBy(header: TableGrid.Header, index: number): void;
-		updateOrderBy(header: TableGrid.Header): void;
-		updateVirtualScroll(): void;
+		clearSelect: ($select: any, $index: number, $event: any) => void;
+		stripHtml: (html: string) => string;
+		updateFilterBy: (header: TableGrid.Header, index: number) => void;
+		updateFilteredAndOrderedRows: () => void;
+		resizedHeaders: () => void;
+		updateOrderBy: (header: TableGrid.Header) => void;
+		updateVirtualScroll: () => void;
+		onMasterCheckBoxChange: () => void;
+		onCheckBoxChange: () => void;
+		getCheckboxState: () => string;
+		refresh: () => void;
 	}
 
 }